requirements.txt
openai==1.42.0
pymupdf==1.24.8
streamlit==1.40.1
python-docx==1.1.2
python-dotenv==1.0.0